CDP: split net/transport code out of lib.rs into a transport module
